Curve Editor: Check for valid key handle before removing
#rb Mike.Zyracki
#jira UE-171677

[CL 35717091 by max chen in ue5-main branch]

+++ Engine/Source/Editor/CurveEditor/Private/RichCurveEditorModel.cpp
@@ -208,7 +208,10 @@ void FRichCurveEditorModel::RemoveKeys(TArrayView<const FKeyHandle> InKeys)
 			FRichCurve& RichCurve = GetRichCurve();
 			for (FKeyHandle Handle : InKeys)
 			{
-				RichCurve.DeleteKey(Handle);
+				if (RichCurve.IsKeyHandleValid(Handle))
+				{
+					RichCurve.DeleteKey(Handle);
+				}
 			}
 
 			CurveModifiedDelegate.Broadcast();
